Extract createToolsSlice from projectStore (P5 store slice)
